Added support for MongoDB backend via Mongoid.
Mongoid passes all backend rspec tests.

Updated gemspect to include mongoid gem for development.

Modified the shared_specs so that they are not nested.  When shared example
groups are nested there is an error that is caused when they are used more than
once.

Added :mongoid as the default backend after ActiveRecord when Mongoid is defined.