Do not misinterpret log errors as job errors
Fix #850

In the event that an exception occurs while the logger is logging the
successful running of the job, DelayedJob will assume that the exception
occurred during the job invocation, and then proceed to handle the job
as though it had failed. If the job succeeded, it's likely to be
destroyed and frozen, and will raise an unhandled exception about not
being able to modify a frozen object.

This exception is confusing. I propose that DelayedJob instead doesn't
rescue log-related failures.