Only perform error callback when job really failed
This fixes #850, which describes a specific scenario where DJ workers
can crash when the logger fails to log that a job was processed
successfully.

The one downside of this fix is that some messages won't be logged, and
the developer won't know about it. This is why I initially implemented
PR #852 to not rescue those logging exceptions, but it sounds like
keeping the workers alive is more important than knowing if a log
failed. That kind of makes sense to me.