Add failing test for Ruby 2.3 + Mongoid
I'm trying to upgrade to Ruby 2.3, and I'm observing some failing tests,
which I've tried to reproduce here for you. This test only fails for
delayed_job_mongoid, not delayed_job_active_record. So the build for
this repo will be green. But I'll make a PR against the other one to
demonstrate the failure.

The issue, as I understand it, is this: when Ruby serializes an object
to yaml and there are multiple references to the same value, it will use
"aliases" to avoid duplicating the data and save space. That's sweet!
But it seems like DJ can't deserialize those aliases when they're
referring to instances of mongoid models.