Allow graceful interruption of worker sleep
Previously, unless `Worker.raise_signal_exceptions` was set, the worker
process ignored INT and TERM signals while sleeping after detection of
an empty queue. That behavior was specially noticeable when
`sleep_delay` was much greater than the default of 5 seconds: a
daemonized job, for example, would end up being killed because of not
responding in a reasonable time.

Now, when the worker needs to sleep, it does so waiting on an event with
a timeout. Stopping the job from a signal sets the event and aborts the
wait immediately.