Add option to exclude locked jobs (i.e. currently being processed) in jobs:check.
If a job is currently being processed, we want an option to exclude them from being reported in `jobs:check`.

`jobs:check`'s primary purpose should be to see the outstanding queue, not determining long-running jobs. 

Made this option default to false so no breaking changes but simply pass `true` for `not_locked` to exclude them from this check.