Fix orphan workers on restart
If 1 daemonized DJ worker is running and you execute restart with a
different n value, the existing worker will effectively be orphaned, in
that it continues to run through future stop/start/restart commands
with the new n value. The opposite is also true. If you change from an
n value of 2+ to 1 when you restart then the running workers will be
orphaned. This fixes that problem by looking in the pid_dir to identify
all worker process names and stopping them ALL prior to starting n new
workers.