Make reloading more thread-safe.
Reloader will determine whether a reload is needed, wait until it's safe to reload, and then do so, before continuing.
If delayed worker is run in the same process as application (but in a different thread) and Reloader is not used, there might be interference between the threads when they reload concurrently: constants might be not properly loaded and unloaded. For instance, a constant that is being loaded in the main thread might be too early detected as a new constant in the delayed worker thread and, therefore, not be added to the list of autoloaded constants at the end of the loading in the main thread, which then leads to the constant not being unloaded, which then leads to such errors as "A copy of ... has been removed from the module tree but is still active!".
Also, using Reloader may save resources since code will be reloaded not every few seconds but only when a reload is required.